attempt to prevent encoder collapse
After a bit of experimentation, it seemed that the encoder outputs were collapsing to the same value. By zero-initializing the last layer of the classifier, scaling input weights in the encoder, and increasing the classifier size, I think I managed to prevent the collapse. In the very least, the collapse takes a little longer now.